+++ tmp-tests/test-new-split2.cpp
@@ -1,5 +1,7 @@
 /******************************************************************************/
 
+#define ARMA_64BIT_WORD
+
 #include <bigstatsr/arma-strict-R-headers.h>
 
 using namespace Rcpp;
@@ -7,11 +9,11 @@ using namespace Rcpp;
 /******************************************************************************/
 
 // [[Rcpp::export]]
-List get_L(std::vector<size_t> p,
-           const IntegerVector& i,
-           const NumericVector& x,
-           double thr_r2,
-           double max_r2) {
+List get_L2(std::vector<size_t> p,
+            const IntegerVector& i,
+            const NumericVector& x,
+            double thr_r2,
+            double max_r2) {
 
   // L(i, j) = sum_{q=j}^m r(i, q)^2 (note: start at j, but j+1 in the paper)
   // L(i, i) is not used in E(i, j), so not computed
@@ -57,10 +59,11 @@ List get_L(std::vector<size_t> p,
 /******************************************************************************/
 
 // [[Rcpp::export]]
-List get_C(const arma::sp_mat& L, int min_size, int max_size, int max_K, double max_cost) {
+List get_C2(const arma::sp_mat& L, int min_size, int max_size, int max_K) {
 
   int m = L.n_rows;  // L now has an extra column with all 0s for convenience
   std::vector< std::vector<float> > res_E(m);
+  double max_E = double(m) * m;
 
   // E(i, j) = sum_{p=i}^j L(p, j+1)
 
@@ -72,12 +75,11 @@ List get_C(const arma::sp_mat& L, int min_size, int max_size, int max_K, double 
     for (int row = col; row >= 0; row--) {
 
       e += L(row, col + 1);  // compute E(j, j), then E(j-1, j), etc
-      if (e > max_cost) break;
+      if (e > max_E) break;
       count++;
 
       if (count >= min_size) {
-        // E(row, col) = e with reparametrization to save memory
-        // -> the first row has to be (col - min_size + 1)
+        // E[row, col] = e with reparametrization to save memory
         res_E[col].push_back(e);
         if (count == max_size) break;
       }
@@ -111,48 +113,9 @@ List get_C(const arma::sp_mat& L, int min_size, int max_size, int max_K, double 
         }
       }
     }
-
-    if (C(0, k) > max_cost && C(0, k) > C(0, k - 1)) break;
   }
 
   return List::create(_["C"] = C, _["best_ind"] = best_ind);
-}
-
-/******************************************************************************/
-
-// [[Rcpp::export]]
-double get_perc(const NumericVector& p,  // integers, but possibly very large
-                const IntegerVector& i,
-                const IntegerVector& all_last) {
-
-  int m = p.size() - 1;
-  double count_all = 2 * i.size() - m;  // count the diaginal only once
-  double count_within = count_all;
-
-  int grp_num = 0;
-  int limit = all_last[grp_num];
-
-  for (int j = 0; j < m; j++) {
-
-    if (j > limit) {
-      grp_num++;
-      limit = all_last[grp_num];
-    }
-
-    size_t lo = p[j];      // the diagonal -> can skip it
-    size_t up = p[j + 1];  // up > lo because of the diag
-
-    // start by the end to stop early
-    for (size_t k = up - 1; k > lo; k--) {
-      if (i[k] > limit) {
-        count_within -= 2;
-      } else {
-        break;  // all remaining i are in the block
-      }
-    }
-  }
-
-  return count_within / count_all;
 }
 
 /******************************************************************************/